moved everything of network usage into its own class
